dashboard/requirements.txt
streamlit
pandas-gbq
google-cloud-bigquery
scikit-learn
joblib
plotly
